requirements.txt
asgiref==3.7.2
Django==4.2.4
django-admin==2.0.2
django-excel-response2==3.0.6
django-rest-framework==0.1.0
django-six==1.0.5
djangorestframework==3.14.0
excel-base==1.0.4
isoweek==1.3.3
python-dateutil==2.8.2
pytz==2023.3
screen==1.0.1
six==1.16.0
sqlparse==0.4.4
TimeConvert==3.0.13
typing_extensions==4.7.1
tzlocal==5.0.1
xlwt==1.3.0
